Remove super user options check box and bug fix (#431)
* chore(i429) remove super user options check box and bug fix

* chore(i429) remove super user options check box and bug fix

* memoize the useEffect functions

* chore(i429) changed margin to padding

* chore(i429) removed userProfile file and its scss file
