+++ src/components/home/index.js
@@ -1,13 +1,12 @@
+import React, { useState, useEffect, useCallback } from 'react';
 import { useRouter } from 'next/router';
 import Members from '@components/members';
 import NewMembers from '@components/new-members';
-import UserProfile from '@components/user-profile';
 import MemberRoleUpdate from '@components/member-role-update';
 import Designers from '@components/designers';
 import SearchBox from '@components/UI/search-box/index';
 import styles from '@components/home/home.module.scss';
 import { userContext } from '@store/user/user-context';
-import { useState } from 'react';
 import {
   BRAND_NAME,
   MEMBERS_TITLE,
@@ -15,10 +14,19 @@ import {
 } from '@constants/AppConstants';
 
 const Home = () => {
-  const { showMemberRoleUpdateModal, isSuperUserMode } = userContext();
+  const { showMemberRoleUpdateModal, isSuperUser, setUserPrivileges } =
+    userContext();
   const { query } = useRouter() || { query: { dev: false } };
   const { dev } = query;
   const [isOptionKey, setIsOptionKey] = useState(false);
+
+  const memoizeUserPrivilege = useCallback(async () => {
+    await setUserPrivileges();
+  }, [setUserPrivileges]);
+
+  useEffect(() => {
+    memoizeUserPrivilege();
+  }, [memoizeUserPrivilege]);
 
   return (
     <div
@@ -34,12 +42,11 @@ const Home = () => {
         setIsOptionKey(false);
       }}
     >
-      <UserProfile />
       {dev && <SearchBox />}
       <h1 className={styles.heading}>Designers</h1>
       <Designers />
       <h1 className={styles.heading}>{`${BRAND_NAME} ${MEMBERS_TITLE}`}</h1>
-      {isSuperUserMode && (
+      {isSuperUser && (
         <div id="memberRoleUpdateModal">
           {showMemberRoleUpdateModal && <MemberRoleUpdate />}
         </div>
