+++ src/components/member-card/card.module.scss
@@ -14,10 +14,6 @@
   height: 50px;
 }
 
-.newUsers {
-  pointer-events: none;
-}
-
 .nameOfPerson {
   padding: 10px 10px 0 10px;
   font-size: 1.43rem;
