+++ src/components/member-card/index.js
@@ -41,7 +41,6 @@ const Card = ({ developerInfo, isOptionKey }) => {
       onMouseLeave={() => {
         setShowSettings(false);
       }}
-      className={isMember ? [] : classNames.newUsers}
     >
       {dev && (
         <SuperUserOptions username={username} showSettings={showSettings} />
