--- src/components/user-profile/user-profile.module.scss
@@ -1,4 +1,0 @@
-.superUserOptions {
-  position: absolute;
-  right: 0;
-}
