+++ src/store/user/user-context.js
@@ -7,7 +7,6 @@ const UserContext = createContext();
 export const UserContextProvider = ({ children }) => {
   const [user, setUser] = useState(null);
   const [isSuperUser, setIsSuperUser] = useState(false);
-  const [isSuperUserMode, setIsSuperUserMode] = useState(false);
   const [showMemberRoleUpdateModal, setShowMemberRoleUpdateModal] =
     useState(false);
   const [selectedMember, setSelectedMember] = useState(null);
@@ -23,12 +22,10 @@ export const UserContextProvider = ({ children }) => {
   const initialUserContext = {
     user,
     isSuperUser,
-    isSuperUserMode,
     selectedMember,
     showMemberRoleUpdateModal,
     setIsSuperUser,
     setUser,
-    setIsSuperUserMode,
     setSelectedMember,
     setShowMemberRoleUpdateModal,
     userApiCalled,
